@@ -17,7 +17,7 @@ def _subword_start_by_joiner(tokens: Sequence[str]) -> Sequence[bool]:
             flag[i] = False
         if token.endswith(SubwordMarker.JOINER):
             try:
-                flag[i+1] = False
+                flag[i + 1] = False
             except IndexError:
                 print("Sentence `{}` not correct!".format(" ".join(token)))
                 raise
@@ -45,11 +45,21 @@ def word_start_finder(ignore_subword=False, is_joiner=False) -> Callable:
 class BARTNoising(object):
     """Noise from BART."""
 
-    def __init__(self, vocab, mask_tok=DefaultTokens.MASK, mask_ratio=0.0,
-                 insert_ratio=0.0, permute_sent_ratio=0.0, poisson_lambda=3.0,
-                 replace_length=-1, rotate_ratio=0.0, mask_length='subword',
-                 random_ratio=0.0, is_joiner=False,
-                 full_stop_token=DefaultTokens.SENT_FULL_STOPS):
+    def __init__(
+        self,
+        vocab,
+        mask_tok=DefaultTokens.MASK,
+        mask_ratio=0.0,
+        insert_ratio=0.0,
+        permute_sent_ratio=0.0,
+        poisson_lambda=3.0,
+        replace_length=-1,
+        rotate_ratio=0.0,
+        mask_length="subword",
+        random_ratio=0.0,
+        is_joiner=False,
+        full_stop_token=DefaultTokens.SENT_FULL_STOPS,
+    ):
         if vocab is None:
             raise ValueError("Inject BART noise requires a valid vocabulary.")
         self.vocab = vocab
@@ -68,22 +78,22 @@ class BARTNoising(object):
         #  0: replace everything (i.e. no mask)
         #  1: 1 mask per span
         if replace_length not in [-1, 0, 1]:
-            raise ValueError(f'invalid arg: replace_length={replace_length}')
+            raise ValueError(f"invalid arg: replace_length={replace_length}")
         self.replace_length = replace_length
 
-        if mask_length not in ['subword', 'word', 'span-poisson']:
-            raise ValueError(f'invalid arg: mask-length={mask_length}')
-        if mask_length == 'subword' and replace_length not in [0, 1]:
-            raise ValueError('if using subwords, use replace-length=1 or 0')
+        if mask_length not in ["subword", "word", "span-poisson"]:
+            raise ValueError(f"invalid arg: mask-length={mask_length}")
+        if mask_length == "subword" and replace_length not in [0, 1]:
+            raise ValueError("if using subwords, use replace-length=1 or 0")
 
-        if mask_length == 'subword' or is_joiner is None:
+        if mask_length == "subword" or is_joiner is None:
             # view each subword as word start / input is word level token
             self._is_word_start = word_start_finder(ignore_subword=True)
         else:
             self._is_word_start = word_start_finder(is_joiner=is_joiner)
 
         self.mask_span_distribution = None
-        if mask_length == 'span-poisson':
+        if mask_length == "span-poisson":
             self.mask_span_distribution = self._make_poisson(poisson_lambda)
         self.mask_length = mask_length
         self.poisson_lambda = poisson_lambda
@@ -102,7 +112,7 @@ class BARTNoising(object):
         for k in range(0, 128):
             ps.append(e_to_the_minus_lambda * lambda_to_the_k / k_factorial)
             lambda_to_the_k *= poisson_lambda
-            k_factorial *= (k + 1)
+            k_factorial *= k + 1
             if ps[-1] < 0.0000001:
                 break
         ps = torch.FloatTensor(ps)
@@ -111,10 +121,7 @@ class BARTNoising(object):
     def _get_sentence_borders(self, tokens):
         """Return lengths of each sentence in the token sequence."""
         full_stops = np.array(
-            [
-                True if token in self.full_stop_token else False
-                for token in tokens
-            ]
+            [True if token in self.full_stop_token else False for token in tokens]
         )
         # Pretend it ends with a full stop so last span is a sentence
         full_stops[-1] = True
@@ -134,15 +141,13 @@ class BARTNoising(object):
 
         substitutions = np.random.permutation(n_sentences)[:n_to_permute]
         ordering = np.arange(0, n_sentences)
-        ordering[substitutions] = substitutions[np.random.permutation(
-            n_to_permute)]
+        ordering[substitutions] = substitutions[np.random.permutation(n_to_permute)]
 
         result = [tok for tok in tokens]
         index = 0
         for i in ordering:
-            sentence = tokens[(sentence_lens[i - 1] if i > 0 else 0):
-                              sentence_lens[i]]
-            result[index:index + len(sentence)] = sentence
+            sentence = tokens[(sentence_lens[i - 1] if i > 0 else 0) : sentence_lens[i]]
+            result[index : index + len(sentence)] = sentence
             index += len(sentence)
         assert len(result) == len(tokens), "Error when permute sentences."
         return result
@@ -155,17 +160,18 @@ class BARTNoising(object):
             return tokens
 
         if self.mask_span_distribution is not None:  # Text (span) Infilling
-            lengths = self.mask_span_distribution.sample(
-                sample_shape=(n_mask,))
+            lengths = self.mask_span_distribution.sample(sample_shape=(n_mask,))
 
             # Make sure we have enough to mask
             cum_length = torch.cumsum(lengths, 0)
             while cum_length[-1] < n_mask:
-                lengths = torch.cat([
-                    lengths,
-                    self.mask_span_distribution.sample(
-                        sample_shape=(n_mask,))
-                ], dim=0)
+                lengths = torch.cat(
+                    [
+                        lengths,
+                        self.mask_span_distribution.sample(sample_shape=(n_mask,)),
+                    ],
+                    dim=0,
+                )
                 cum_length = torch.cumsum(lengths, 0)
 
             # Trim to masking budget
@@ -188,9 +194,7 @@ class BARTNoising(object):
             lengths = torch.ones((n_mask,)).long()
         # assert is_word_start[-1] == 0
         word_starts = is_word_start.nonzero(as_tuple=False)
-        indices = word_starts[
-            torch.randperm(word_starts.size(0))[:n_mask]
-        ].squeeze(1)
+        indices = word_starts[torch.randperm(word_starts.size(0))[:n_mask]].squeeze(1)
         mask_random = torch.FloatTensor(n_mask).uniform_() < self.random_ratio
 
         tokens_length = len(tokens)
@@ -204,12 +208,13 @@ class BARTNoising(object):
             for i in indices.tolist():
                 tokens[i] = self.mask_tok
             random_tok_ids = torch.randint(
-                0, len(self.vocab), size=(mask_random.sum(),)).tolist()
+                0, len(self.vocab), size=(mask_random.sum(),)
+            ).tolist()
             for i, rid in zip(indices[mask_random].tolist(), random_tok_ids):
                 tokens[i] = self.vocab[rid]
 
         if tokens_length - 1 in indices:
-            uncompleted = (indices != tokens_length - 1)
+            uncompleted = indices != tokens_length - 1
             indices = indices[uncompleted]
             mask_random = mask_random[uncompleted]
             lengths = lengths[uncompleted]
@@ -237,9 +242,9 @@ class BARTNoising(object):
                     for i in indices.tolist():
                         tokens[i] = self.mask_tok
                     random_tok_ids = torch.randint(
-                        0, len(self.vocab), size=(mask_random.sum(),)).tolist()
-                    for i, rid in zip(
-                            indices[mask_random].tolist(), random_tok_ids):
+                        0, len(self.vocab), size=(mask_random.sum(),)
+                    ).tolist()
+                    for i, rid in zip(indices[mask_random].tolist(), random_tok_ids):
                         tokens[i] = self.vocab[rid]
         else:
             # A bit faster when all lengths are 1
@@ -256,15 +261,14 @@ class BARTNoising(object):
                     for i in indices.tolist():
                         tokens[i] = self.mask_tok
                     random_tok_ids = torch.randint(
-                        0, len(self.vocab), size=(mask_random.sum(),)).tolist()
-                    for i, rid in zip(
-                            indices[mask_random].tolist(), random_tok_ids):
+                        0, len(self.vocab), size=(mask_random.sum(),)
+                    ).tolist()
+                    for i, rid in zip(indices[mask_random].tolist(), random_tok_ids):
                         tokens[i] = self.vocab[rid]
 
                 # assert tokens_length - 1 not in indices
 
-        tokens = [tok for tok, keep in zip(tokens, to_keep.tolist())
-                  if keep is True]
+        tokens = [tok for tok, keep in zip(tokens, to_keep.tolist()) if keep is True]
 
         if n_insert > 0:
             tokens = self.insertion_noise(tokens, n_insert / len(tokens))
@@ -286,11 +290,11 @@ class BARTNoising(object):
         result[noise_indices[n_random:]] = self.mask_tok
         if n_random > 0:
             result[noise_indices[:n_random]] = np.random.choice(
-                self.vocab, size=n_random)
+                self.vocab, size=n_random
+            )
         result[~noise_mask] = tokens
 
-        assert all([item is not None for item in result]),\
-            "Error when inserting noise."
+        assert all([item is not None for item in result]), "Error when inserting noise."
         return result.tolist()
 
     def rolling_noise(self, tokens, p=1.0):
@@ -317,25 +321,24 @@ class BARTNoising(object):
         cls_name = type(self).__name__
         kwargs = {}
         if self.permute_sent_ratio > 0.0:
-            kwargs['permute_sent_ratio'] = self.permute_sent_ratio
-            kwargs['full_stop_token'] = self.full_stop_token
+            kwargs["permute_sent_ratio"] = self.permute_sent_ratio
+            kwargs["full_stop_token"] = self.full_stop_token
         if self.insert_ratio > 0.0:
-            kwargs['insert_ratio'] = self.insert_ratio
+            kwargs["insert_ratio"] = self.insert_ratio
         if self.rotate_ratio > 0.0:
-            kwargs['rotate_ratio'] = self.rotate_ratio
+            kwargs["rotate_ratio"] = self.rotate_ratio
         if self.random_ratio > 0.0:
-            kwargs['random_ratio'] = self.random_ratio
+            kwargs["random_ratio"] = self.random_ratio
         if self.mask_ratio > 0.0:
-            kwargs['mask_ratio'] = self.mask_ratio
-            kwargs['mask_length'] = self.mask_length
-            kwargs['poisson_lambda'] = self.poisson_lambda
-            kwargs['replace_length'] = self.replace_length
-        cls_args = ', '.join(
-            [f'{kw}={arg}' for kw, arg in kwargs.items()])
-        return '{}({})'.format(cls_name, cls_args)
+            kwargs["mask_ratio"] = self.mask_ratio
+            kwargs["mask_length"] = self.mask_length
+            kwargs["poisson_lambda"] = self.poisson_lambda
+            kwargs["replace_length"] = self.replace_length
+        cls_args = ", ".join([f"{kw}={arg}" for kw, arg in kwargs.items()])
+        return "{}({})".format(cls_name, cls_args)
 
 
-@register_transform(name='bart')
+@register_transform(name="bart")
 class BARTNoiseTransform(Transform):
     def __init__(self, opts):
         super().__init__(opts)
@@ -348,32 +351,71 @@ class BARTNoiseTransform(Transform):
     def add_options(cls, parser):
         """Avalilable options relate to BART."""
         group = parser.add_argument_group("Transform/BART")
-        group.add("--permute_sent_ratio", "-permute_sent_ratio",
-                  type=float, default=0.0,
-                  help="Permute this proportion of sentences "
-                       "(boundaries defined by {}) in all inputs.".format(
-                        DefaultTokens.SENT_FULL_STOPS))
-        group.add("--rotate_ratio", "-rotate_ratio", type=float, default=0.0,
-                  help="Rotate this proportion of inputs.")
-        group.add("--insert_ratio", "-insert_ratio", type=float, default=0.0,
-                  help="Insert this percentage of additional random tokens.")
-        group.add("--random_ratio", "-random_ratio", type=float, default=0.0,
-                  help="Instead of using {}, use random token "
-                       "this often.".format(DefaultTokens.MASK))
+        group.add(
+            "--permute_sent_ratio",
+            "-permute_sent_ratio",
+            type=float,
+            default=0.0,
+            help="Permute this proportion of sentences "
+            "(boundaries defined by {}) in all inputs.".format(
+                DefaultTokens.SENT_FULL_STOPS
+            ),
+        )
+        group.add(
+            "--rotate_ratio",
+            "-rotate_ratio",
+            type=float,
+            default=0.0,
+            help="Rotate this proportion of inputs.",
+        )
+        group.add(
+            "--insert_ratio",
+            "-insert_ratio",
+            type=float,
+            default=0.0,
+            help="Insert this percentage of additional random tokens.",
+        )
+        group.add(
+            "--random_ratio",
+            "-random_ratio",
+            type=float,
+            default=0.0,
+            help="Instead of using {}, use random token "
+            "this often.".format(DefaultTokens.MASK),
+        )
 
-        group.add("--mask_ratio", "-mask_ratio", type=float, default=0.0,
-                  help="Fraction of words/subwords that will be masked.")
-        group.add("--mask_length", "-mask_length", type=str, default="subword",
-                  choices=["subword", "word", "span-poisson"],
-                  help="Length of masking window to apply.")
-        group.add("--poisson_lambda", "-poisson_lambda",
-                  type=float, default=3.0,
-                  help="Lambda for Poisson distribution to sample span length "
-                       "if `-mask_length` set to span-poisson.")
-        group.add("--replace_length", "-replace_length",
-                  type=int, default=-1, choices=[-1, 0, 1],
-                  help="When masking N tokens, replace with 0, 1, "
-                       "or N tokens. (use -1 for N)")
+        group.add(
+            "--mask_ratio",
+            "-mask_ratio",
+            type=float,
+            default=0.0,
+            help="Fraction of words/subwords that will be masked.",
+        )
+        group.add(
+            "--mask_length",
+            "-mask_length",
+            type=str,
+            default="subword",
+            choices=["subword", "word", "span-poisson"],
+            help="Length of masking window to apply.",
+        )
+        group.add(
+            "--poisson_lambda",
+            "-poisson_lambda",
+            type=float,
+            default=3.0,
+            help="Lambda for Poisson distribution to sample span length "
+            "if `-mask_length` set to span-poisson.",
+        )
+        group.add(
+            "--replace_length",
+            "-replace_length",
+            type=int,
+            default=-1,
+            choices=[-1, 0, 1],
+            help="When masking N tokens, replace with 0, 1, "
+            "or N tokens. (use -1 for N)",
+        )
 
     @classmethod
     def require_vocab(cls):
@@ -384,14 +426,15 @@ class BARTNoiseTransform(Transform):
         super().warm_up(vocabs)
 
         subword_type = self.opts.src_subword_type
-        if self.opts.mask_length == 'subword':
-            if subword_type == 'none':
+        if self.opts.mask_length == "subword":
+            if subword_type == "none":
                 raise ValueError(
-                    f'src_subword_type={subword_type} incompatible with '
-                    f'mask_length={self.opts.mask_length}!')
-        is_joiner = (subword_type == 'bpe') if subword_type != 'none' else None
+                    f"src_subword_type={subword_type} incompatible with "
+                    f"mask_length={self.opts.mask_length}!"
+                )
+        is_joiner = (subword_type == "bpe") if subword_type != "none" else None
         self.bart_noise = BARTNoising(
-            self.vocabs['src'].ids_to_tokens,
+            self.vocabs["src"].ids_to_tokens,
             mask_tok=DefaultTokens.MASK,
             mask_ratio=self.opts.mask_ratio,
             insert_ratio=self.opts.insert_ratio,
@@ -401,14 +444,14 @@ class BARTNoiseTransform(Transform):
             rotate_ratio=self.opts.rotate_ratio,
             mask_length=self.opts.mask_length,
             random_ratio=self.opts.random_ratio,
-            is_joiner=is_joiner
+            is_joiner=is_joiner,
         )
 
     def apply(self, example, is_train=False, stats=None, **kwargs):
         """Apply BART noise to src side tokens."""
         if is_train:
-            src = self.bart_noise.apply(example['src'])
-            example['src'] = src
+            src = self.bart_noise.apply(example["src"])
+            example["src"] = src
         return example
 
     def _repr_args(self):
