@@ -5,14 +5,20 @@ import onmt.decoders
 import onmt.models
 import onmt.utils
 import onmt.modules
-from onmt.trainer import Trainer
 import sys
 import onmt.utils.optimizers
+
 onmt.utils.optimizers.Optim = onmt.utils.optimizers.Optimizer
 sys.modules["onmt.Optim"] = onmt.utils.optimizers
 
 # For Flake
-__all__ = [onmt.inputters, onmt.encoders, onmt.decoders, onmt.models,
-           onmt.utils, onmt.modules, "Trainer"]
+__all__ = [
+    onmt.inputters,
+    onmt.encoders,
+    onmt.decoders,
+    onmt.models,
+    onmt.utils,
+    onmt.modules,
+]
 
-__version__ = "3.0.0"
+__version__ = "3.5.0"
