@@ -2,7 +2,7 @@
 """Train models with dynamic data."""
 import torch
 from functools import partial
-from onmt.utils.distributed import ErrorHandler, consumer
+from onmt.utils.distributed import ErrorHandler, spawned_train
 from onmt.utils.misc import set_random_seed
 from onmt.utils.logging import init_logger, logger
 from onmt.utils.parse import ArgumentParser
@@ -15,7 +15,6 @@ from onmt.train_single import main as single_main
 
 
 def train(opt):
-
     init_logger(opt.log_file)
 
     ArgumentParser.validate_train_opts(opt)
@@ -24,22 +23,25 @@ def train(opt):
 
     set_random_seed(opt.seed, False)
 
-    train_process = partial(
-        single_main)
+    train_process = partial(single_main)
 
     nb_gpu = len(opt.gpu_ranks)
 
     if opt.world_size > 1:
-        mp = torch.multiprocessing.get_context('spawn')
+        mp = torch.multiprocessing.get_context("spawn")
         # Create a thread to listen for errors in the child processes.
         error_queue = mp.SimpleQueue()
         error_handler = ErrorHandler(error_queue)
         # Train with multiprocessing.
         procs = []
         for device_id in range(nb_gpu):
-            procs.append(mp.Process(target=consumer, args=(
-                train_process, opt, device_id, error_queue),
-                daemon=False))
+            procs.append(
+                mp.Process(
+                    target=spawned_train,
+                    args=(train_process, opt, device_id, error_queue),
+                    daemon=False,
+                )
+            )
             procs[device_id].start()
             logger.info(" Starting process pid: %d  " % procs[device_id].pid)
             error_handler.add_child(procs[device_id].pid)
@@ -48,12 +50,12 @@ def train(opt):
 
     elif nb_gpu == 1:  # case 1 GPU only
         train_process(opt, device_id=0)
-    else:   # case only CPU
+    else:  # case only CPU
         train_process(opt, device_id=-1)
 
 
 def _get_parser():
-    parser = ArgumentParser(description='train.py')
+    parser = ArgumentParser(description="train.py")
     train_opts(parser)
     return parser
 
