@@ -16,6 +16,7 @@ import onmt.model_builder
 
 class EnsembleDecoderOutput(object):
     """Wrapper around multiple decoder final hidden states."""
+
     def __init__(self, model_dec_outs):
         self.model_dec_outs = tuple(model_dec_outs)
 
@@ -23,8 +24,7 @@ class EnsembleDecoderOutput(object):
         """Delegate squeeze to avoid modifying
         :func:`onmt.translate.translator.Translator.translate_batch()`
         """
-        return EnsembleDecoderOutput([
-            x.squeeze(dim) for x in self.model_dec_outs])
+        return EnsembleDecoderOutput([x.squeeze(dim) for x in self.model_dec_outs])
 
     def __getitem__(self, index):
         return self.model_dec_outs[index]
@@ -32,49 +32,55 @@ class EnsembleDecoderOutput(object):
 
 class EnsembleEncoder(EncoderBase):
     """Dummy Encoder that delegates to individual real Encoders."""
+
     def __init__(self, model_encoders):
         super(EnsembleEncoder, self).__init__()
         self.model_encoders = nn.ModuleList(model_encoders)
 
     def forward(self, src, src_len=None):
-        enc_out, enc_final_hs, _ = zip(*[
-            model_encoder(src, src_len)
-            for model_encoder in self.model_encoders])
+        enc_out, enc_final_hs, _ = zip(
+            *[model_encoder(src, src_len) for model_encoder in self.model_encoders]
+        )
         return enc_out, enc_final_hs, src_len
 
 
 class EnsembleDecoder(DecoderBase):
     """Dummy Decoder that delegates to individual real Decoders."""
+
     def __init__(self, model_decoders):
         model_decoders = nn.ModuleList(model_decoders)
         attentional = any([dec.attentional for dec in model_decoders])
         super(EnsembleDecoder, self).__init__(attentional)
         self.model_decoders = model_decoders
 
-    def forward(self, tgt, enc_out, src_len=None, step=None,
-                **kwargs):
+    def forward(self, tgt, enc_out, src_len=None, step=None, **kwargs):
         """See :func:`onmt.decoders.decoder.DecoderBase.forward()`."""
         # src_len is a single tensor shared between all models.
         # This assumption will not hold if Translator is modified
         # to calculate src_len as something other than the length
         # of the input.
-        dec_outs, attns = zip(*[
-            model_decoder(
-                tgt, enc_out[i],
-                src_len=src_len, step=step, **kwargs)
-            for i, model_decoder in enumerate(self.model_decoders)])
-        mean_attns = self.combine_attns(attns)
+        dec_outs, attns = zip(
+            *[
+                model_decoder(tgt, enc_out[i], src_len=src_len, step=step, **kwargs)
+                for i, model_decoder in enumerate(self.model_decoders)
+            ]
+        )
+        if attns[0]["std"] is not None:
+            mean_attns = self.combine_attns(attns)
+        else:
+            mean_attns = attns
         return EnsembleDecoderOutput(dec_outs), mean_attns
 
     def combine_attns(self, attns):
         result = {}
         for key in attns[0].keys():
             result[key] = torch.stack(
-                [attn[key] for attn in attns if attn[key] is not None]).mean(0)
+                [attn[key] for attn in attns if attn[key] is not None]
+            ).mean(0)
         return result
 
     def init_state(self, src, enc_out, enc_hidden):
-        """ See :obj:`RNNDecoderBase.init_state()` """
+        """See :obj:`RNNDecoderBase.init_state()`"""
         for i, model_decoder in enumerate(self.model_decoders):
             model_decoder.init_state(src, enc_out[i], enc_hidden[i])
 
@@ -88,6 +94,7 @@ class EnsembleGenerator(nn.Module):
     Dummy Generator that delegates to individual real Generators,
     and then averages the resulting target distributions.
     """
+
     def __init__(self, model_generators, raw_probs=False):
         super(EnsembleGenerator, self).__init__()
         self.model_generators = nn.ModuleList(model_generators)
@@ -100,9 +107,11 @@ class EnsembleGenerator(nn.Module):
         All models in the ensemble must share a target vocabulary.
         """
         distributions = torch.stack(
-                [mg(h) if attn is None else mg(h, attn, src_map)
-                 for h, mg in zip(hidden, self.model_generators)]
-            )
+            [
+                mg(h) if attn is None else mg(h, attn, src_map)
+                for h, mg in zip(hidden, self.model_generators)
+            ]
+        )
         if self._raw_probs:
             return torch.log(torch.exp(distributions).mean(0))
         else:
@@ -111,29 +120,32 @@ class EnsembleGenerator(nn.Module):
 
 class EnsembleModel(NMTModel):
     """Dummy NMTModel wrapping individual real NMTModels."""
+
     def __init__(self, models, raw_probs=False):
         encoder = EnsembleEncoder(model.encoder for model in models)
         decoder = EnsembleDecoder(model.decoder for model in models)
         super(EnsembleModel, self).__init__(encoder, decoder)
         self.generator = EnsembleGenerator(
-            [model.generator for model in models], raw_probs)
+            [model.generator for model in models], raw_probs
+        )
         self.models = nn.ModuleList(models)
 
 
-def load_test_model(opt):
+def load_test_model(opt, device_id=0):
     """Read in multiple models for ensemble."""
     shared_vocabs = None
     shared_model_opt = None
     models = []
     for model_path in opt.models:
-        vocabs, model, model_opt = \
-            onmt.model_builder.load_test_model(opt, model_path=model_path)
+        vocabs, model, model_opt = onmt.model_builder.load_test_model(
+            opt, device_id, model_path=model_path
+        )
         if shared_vocabs is None:
             shared_vocabs = vocabs
         else:
-            assert shared_vocabs['src'].tokens_to_ids == \
-                vocabs['src'].tokens_to_ids, \
-                "Ensemble models must use the same vocabs "
+            assert (
+                shared_vocabs["src"].tokens_to_ids == vocabs["src"].tokens_to_ids
+            ), "Ensemble models must use the same vocabs "
         models.append(model)
         if shared_model_opt is None:
             shared_model_opt = model_opt
