@@ -1,5 +1,5 @@
 """Module defining models."""
 from onmt.models.model_saver import build_model_saver, ModelSaver
-from onmt.models.model import NMTModel
+from onmt.models.model import BaseModel, NMTModel, LanguageModel
 
-__all__ = ["build_model_saver", "ModelSaver", "NMTModel"]
+__all__ = ["build_model_saver", "ModelSaver", "BaseModel", "NMTModel", "LanguageModel"]
