@@ -51,7 +51,50 @@ class TestServerModel(unittest.TestCase):
         with self.assertRaises(ValueError):
             sm.detokenize("hello world")
 
+    def test_source_features(self):
+        model_id = 0
+        opt = {"models": ["test_model.pt"]}
+        model_root = TEST_DIR
+        sm = ServerModel(
+            opt,
+            model_id,
+            model_root=model_root,
+            load=True,
+            features_opt={
+                "n_src_feats": 1,
+                "src_feats_defaults": "0",
+                "reversible_tokenization": "joiner",
+            },
+        )
+        feats = sm.maybe_transform_feats("hello world.", "hello world ￭.", ["0 1"])
+        self.assertEqual(feats, ["0 1 1"])
+        preprocessed = sm.maybe_preprocess({"src": "hello￨0 world￨1"})
+        self.assertEqual(preprocessed["seg"], ["hello world"])
+        self.assertEqual(preprocessed["src_feats"], [["0 1"]])
+
+    def test_passing_source_features_without_proper_configuration(self):
+        model_id = 0
+        opt = {"models": ["test_model.pt"]}
+        model_root = TEST_DIR
+        sm = ServerModel(opt, model_id, model_root=model_root, load=True)
+        with self.assertRaises(AssertionError):
+            sm.maybe_preprocess({"src": "hello￨0 world￨1"})
+        sm = ServerModel(
+            opt,
+            model_id,
+            model_root=model_root,
+            load=True,
+            features_opt={
+                "n_src_feats": 2,
+                "src_feats_defaults": "0￨0",
+                "reversible_tokenization": "joiner",
+            },
+        )
+        with self.assertRaises(AssertionError):
+            sm.maybe_preprocess({"src": "hello￨0 world￨1"})
+
     if torch.cuda.is_available():
+
         def test_moving_to_gpu_and_back(self):
             torch.cuda.set_device(torch.device("cuda", 0))
             model_id = 0
@@ -86,13 +129,13 @@ class TestServerModel(unittest.TestCase):
                 self.assertEqual(p.device.type, "cpu")
 
         if torch.cuda.device_count() > 1:
+
             def test_initialize_on_nonzero_gpu_and_back(self):
                 torch.cuda.set_device(torch.device("cuda", 1))
                 model_id = 0
                 opt = {"models": ["test_model.pt"], "gpu": 1}
                 model_root = TEST_DIR
-                sm = ServerModel(opt, model_id, model_root=model_root,
-                                 load=True)
+                sm = ServerModel(opt, model_id, model_root=model_root, load=True)
                 for p in sm.translator.model.parameters():
                     self.assertEqual(p.device.type, "cuda")
                     self.assertEqual(p.device.index, 1)
@@ -109,9 +152,8 @@ class TestServerModel(unittest.TestCase):
         opt = {"models": ["test_model.pt"]}
         model_root = TEST_DIR
         sm = ServerModel(opt, model_id, model_root=model_root, load=True)
-        inp = [{"src": "hello how are you today"},
-               {"src": "good morning to you ."}]
-        results, scores, n_best, time, aligns = sm.run(inp)
+        inp = [{"src": "hello how are you today"}, {"src": "good morning to you ."}]
+        results, scores, n_best, time, aligns, align_scores = sm.run(inp)
         self.assertIsInstance(results, list)
         for sentence_string in results:
             self.assertIsInstance(sentence_string, str)
@@ -122,6 +164,10 @@ class TestServerModel(unittest.TestCase):
         for align_list in aligns:
             for align_string in align_list:
                 if align_string is not None:
+                    self.assertIsInstance(align_string, str)
+        for align_scores_list in align_scores:
+            for score_string in align_scores_list:
+                if score_string is not None:
                     self.assertIsInstance(align_string, str)
         self.assertEqual(len(results), len(scores))
         self.assertEqual(len(scores), len(inp) * n_best)
@@ -134,8 +180,7 @@ class TestTranslationServer(unittest.TestCase):
     # this could be considered an integration test because it touches
     # the filesystem for the config file (and the models)
 
-    CFG_F = os.path.join(
-        TEST_DIR, "test_translation_server_config_file.json")
+    CFG_F = os.path.join(TEST_DIR, "test_translation_server_config_file.json")
 
     def tearDown(self):
         if os.path.exists(self.CFG_F):
@@ -145,7 +190,8 @@ class TestTranslationServer(unittest.TestCase):
         with open(self.CFG_F, "w") as f:
             f.write(cfg)
 
-    CFG_NO_LOAD = dedent("""\
+    CFG_NO_LOAD = dedent(
+        """\
         {
             "models_root": "%s",
             "models": [
@@ -161,7 +207,9 @@ class TestTranslationServer(unittest.TestCase):
                 }
             ]
         }
-        """ % TEST_DIR)
+        """
+        % TEST_DIR
+    )
 
     def test_start_without_initial_loading(self):
         self.write(self.CFG_NO_LOAD)
@@ -170,7 +218,8 @@ class TestTranslationServer(unittest.TestCase):
         self.assertFalse(sv.models[100].loaded)
         self.assertEqual(set(sv.models.keys()), {100})
 
-    CFG_LOAD = dedent("""\
+    CFG_LOAD = dedent(
+        """\
         {
             "models_root": "%s",
             "models": [
@@ -186,7 +235,9 @@ class TestTranslationServer(unittest.TestCase):
                 }
             ]
         }
-        """ % TEST_DIR)
+        """
+        % TEST_DIR
+    )
 
     def test_start_with_initial_loading(self):
         self.write(self.CFG_LOAD)
@@ -195,7 +246,8 @@ class TestTranslationServer(unittest.TestCase):
         self.assertTrue(sv.models[100].loaded)
         self.assertEqual(set(sv.models.keys()), {100})
 
-    CFG_2_MODELS = dedent("""\
+    CFG_2_MODELS = dedent(
+        """\
         {
             "models_root": "%s",
             "models": [
@@ -221,7 +273,9 @@ class TestTranslationServer(unittest.TestCase):
                 }
             ]
         }
-        """ % TEST_DIR)
+        """
+        % TEST_DIR
+    )
 
     def test_start_with_two_models(self):
         self.write(self.CFG_2_MODELS)
