@@ -1,14 +1,25 @@
 """ Modules for translation """
-from onmt.translate.translator import Translator
+from onmt.translate.translator import Translator, GeneratorLM
 from onmt.translate.translation import Translation, TranslationBuilder
 from onmt.translate.beam_search import BeamSearch, GNMTGlobalScorer
+from onmt.translate.beam_search import BeamSearchLM
 from onmt.translate.decode_strategy import DecodeStrategy
-from onmt.translate.greedy_search import GreedySearch
+from onmt.translate.greedy_search import GreedySearch, GreedySearchLM
 from onmt.translate.penalties import PenaltyBuilder
-from onmt.translate.translation_server import TranslationServer, \
-    ServerModelError
+from onmt.translate.translation_server import TranslationServer, ServerModelError
 
-__all__ = ['Translator', 'Translation', 'BeamSearch',
-           'GNMTGlobalScorer', 'TranslationBuilder',
-           'PenaltyBuilder', 'TranslationServer', 'ServerModelError',
-           "DecodeStrategy", "GreedySearch"]
+__all__ = [
+    "Translator",
+    "Translation",
+    "BeamSearch",
+    "GNMTGlobalScorer",
+    "TranslationBuilder",
+    "PenaltyBuilder",
+    "TranslationServer",
+    "ServerModelError",
+    "DecodeStrategy",
+    "GreedySearch",
+    "GreedySearchLM",
+    "BeamSearchLM",
+    "GeneratorLM",
+]
