@@ -46,49 +46,32 @@ def check_path(path, exist_ok=False, log=print):
         else:
             raise IOError(f"path {path} exists, stop.")
     else:
-        os.makedirs(os.path.dirname(path), exist_ok=True)
+        if os.path.dirname(path) != "":
+            os.makedirs(os.path.dirname(path), exist_ok=True)
 
 
 def sequence_mask(lengths, max_len=None):
     """
     Creates a boolean mask from sequence lengths.
     """
-    batch_size = lengths.numel()
     max_len = max_len or lengths.max()
-    return (torch.arange(0, max_len, device=lengths.device)
-            .type_as(lengths)
-            .repeat(batch_size, 1)
-            .lt(lengths.unsqueeze(1)))
+    return torch.arange(0, max_len, device=lengths.device) >= lengths.unsqueeze(1)
 
 
 def tile(x, count, dim=0):
     """
     Tiles x on dimension dim count times.
     """
-    perm = list(range(len(x.size())))
-    if dim != 0:
-        perm[0], perm[dim] = perm[dim], perm[0]
-        x = x.permute(perm)
-    out_size = list(x.size())
-    out_size[0] *= count
-    batch = x.size(0)
-    x = x.contiguous().view(batch, -1) \
-         .transpose(0, 1) \
-         .repeat(count, 1) \
-         .transpose(0, 1) \
-         .contiguous() \
-         .view(*out_size)
-    if dim != 0:
-        x = x.permute(perm).contiguous()
-    return x
+    return x.repeat_interleave(count, dim)
 
 
 def use_gpu(opt):
     """
     Creates a boolean if gpu used
     """
-    return (hasattr(opt, 'gpu_ranks') and len(opt.gpu_ranks) > 0) or \
-        (hasattr(opt, 'gpu') and opt.gpu > -1)
+    return (hasattr(opt, "gpu_ranks") and len(opt.gpu_ranks) > 0) or (
+        hasattr(opt, "gpu") and opt.gpu > -1
+    )
 
 
 def set_random_seed(seed, is_cuda):
@@ -116,16 +99,10 @@ def fn_args(fun):
 
 def report_matrix(row_label, column_label, matrix):
     header_format = "{:>10.10} " + "{:>10.7} " * len(row_label)
-    row_format = "{:>10.10} " + "{:>10.7f} " * len(row_label)
-    output = header_format.format("", *row_label) + '\n'
+    output = header_format.format("", *row_label) + "\n"
     for word, row in zip(column_label, matrix):
-        max_index = row.index(max(row))
-        row_format = row_format.replace(
-            "{:>10.7f} ", "{:*>10.7f} ", max_index + 1)
-        row_format = row_format.replace(
-            "{:*>10.7f} ", "{:>10.7f} ", max_index)
-        output += row_format.format(word, *row) + '\n'
-        row_format = "{:>10.10} " + "{:>10.7f} " * len(row_label)
+        row_format = "{:>10.10} " + "{:>10.7f} " * len(row)
+        output += row_format.format(word, *row) + "\n"
     return output
 
 
@@ -135,8 +112,8 @@ def check_model_config(model_config, root):
         model_path = os.path.join(root, model)
         if not os.path.exists(model_path):
             raise FileNotFoundError(
-                "{} from model {} does not exist".format(
-                    model_path, model_config["id"]))
+                "{} from model {} does not exist".format(model_path, model_config["id"])
+            )
     if "tokenizer" in model_config.keys():
         if "params" in model_config["tokenizer"].keys():
             for k, v in model_config["tokenizer"]["params"].items():
@@ -145,4 +122,6 @@ def check_model_config(model_config, root):
                     if not os.path.exists(tok_path):
                         raise FileNotFoundError(
                             "{} from model {} does not exist".format(
-                                tok_path, model_config["id"]))
+                                tok_path, model_config["id"]
+                            )
+                        )
