@@ -10,17 +10,16 @@ from onmt.utils.logging import logger
 def build_report_manager(opt, gpu_rank):
     if opt.tensorboard and gpu_rank <= 0:
         from torch.utils.tensorboard import SummaryWriter
-        if not hasattr(opt, 'tensorboard_log_dir_dated'):
+
+        if not hasattr(opt, "tensorboard_log_dir_dated"):
             opt.tensorboard_log_dir_dated = (
-                opt.tensorboard_log_dir +
-                datetime.now().strftime("/%b-%d_%H-%M-%S")
+                opt.tensorboard_log_dir + datetime.now().strftime("/%b-%d_%H-%M-%S")
             )
         writer = SummaryWriter(opt.tensorboard_log_dir_dated, comment="Unmt")
     else:
         writer = None
 
-    report_mgr = ReportMgr(opt.report_every, start_time=-1,
-                           tensorboard_writer=writer)
+    report_mgr = ReportMgr(opt.report_every, start_time=-1, tensorboard_writer=writer)
     return report_mgr
 
 
@@ -32,7 +31,7 @@ class ReportMgrBase(object):
         * `_report_step`
     """
 
-    def __init__(self, report_every, start_time=-1.):
+    def __init__(self, report_every, start_time=-1.0):
         """
         Args:
             report_every(int): Report status every this many sentences
@@ -48,12 +47,12 @@ class ReportMgrBase(object):
     def log(self, *args, **kwargs):
         logger.info(*args, **kwargs)
 
-    def report_training(self, step, num_steps, learning_rate, patience,
-                        report_stats, multigpu=False):
+    def report_training(
+        self, step, num_steps, learning_rate, patience, report_stats, multigpu=False
+    ):
         """
         This is the user-defined batch-level traing progress
         report function.
-
         Args:
             step(int): current step count.
             num_steps(int): total number of batches.
@@ -63,25 +62,26 @@ class ReportMgrBase(object):
             report_stats(Statistics): updated Statistics instance.
         """
         if self.start_time < 0:
-            raise ValueError("""ReportMgr needs to be started
-                                (set 'start_time' or use 'start()'""")
+            raise ValueError(
+                """ReportMgr needs to be started
+                                (set 'start_time' or use 'start()'"""
+            )
 
         if step % self.report_every == 0:
             if multigpu:
-                report_stats = \
-                    onmt.utils.Statistics.all_gather_stats(report_stats)
+                report_stats = onmt.utils.Statistics.all_gather_stats(report_stats)
             self._report_training(
-                step, num_steps, learning_rate, patience, report_stats)
+                step, num_steps, learning_rate, patience, report_stats
+            )
             return onmt.utils.Statistics()
         else:
             return report_stats
 
     def _report_training(self, *args, **kwargs):
-        """ To be overridden """
+        """To be overridden"""
         raise NotImplementedError()
 
-    def report_step(self, lr, patience, step, train_stats=None,
-                    valid_stats=None):
+    def report_step(self, lr, patience, step, train_stats=None, valid_stats=None):
         """
         Report stats of a step
 
@@ -93,16 +93,15 @@ class ReportMgrBase(object):
             valid_stats(Statistics): validation stats
         """
         self._report_step(
-            lr, patience, step,
-            train_stats=train_stats,
-            valid_stats=valid_stats)
+            lr, patience, step, valid_stats=valid_stats, train_stats=train_stats
+        )
 
     def _report_step(self, *args, **kwargs):
         raise NotImplementedError()
 
 
 class ReportMgr(ReportMgrBase):
-    def __init__(self, report_every, start_time=-1., tensorboard_writer=None):
+    def __init__(self, report_every, start_time=-1.0, tensorboard_writer=None):
         """
         A report manager that writes statistics on standard output as well as
         (optionally) TensorBoard
@@ -115,56 +114,44 @@ class ReportMgr(ReportMgrBase):
         super(ReportMgr, self).__init__(report_every, start_time)
         self.tensorboard_writer = tensorboard_writer
 
-    def maybe_log_tensorboard(self, stats, prefix, learning_rate,
-                              patience, step):
+    def maybe_log_tensorboard(self, stats, prefix, learning_rate, patience, step):
         if self.tensorboard_writer is not None:
             stats.log_tensorboard(
-                prefix, self.tensorboard_writer, learning_rate, patience, step)
+                prefix, self.tensorboard_writer, learning_rate, patience, step
+            )
 
-    def _report_training(self, step, num_steps, learning_rate, patience,
-                         report_stats):
+    def _report_training(self, step, num_steps, learning_rate, patience, report_stats):
         """
         See base class method `ReportMgrBase.report_training`.
         """
-        report_stats.output(step, num_steps,
-                            learning_rate, self.start_time)
-
-        self.maybe_log_tensorboard(report_stats,
-                                   "progress",
-                                   learning_rate,
-                                   patience,
-                                   step)
+        report_stats.output(step, num_steps, learning_rate, self.start_time)
+        self.maybe_log_tensorboard(
+            report_stats, "progress", learning_rate, patience, step
+        )
         report_stats = onmt.utils.Statistics()
 
         return report_stats
 
-    def _report_step(self, lr, patience, step,
-                     train_stats=None,
-                     valid_stats=None):
+    def _report_step(self, lr, patience, step, valid_stats=None, train_stats=None):
         """
         See base class method `ReportMgrBase.report_step`.
         """
         if train_stats is not None:
-            self.log('Train perplexity: %g' % train_stats.ppl())
-            self.log('Train accuracy: %g' % train_stats.accuracy())
-            self.log('Sentences processed: %g' % train_stats.n_sents)
-            self.log('Average bsz: %4.0f/%4.0f/%2.0f' %
-                     (train_stats.n_src_words / train_stats.n_batchs,
-                      train_stats.n_words / train_stats.n_batchs,
-                      train_stats.n_sents / train_stats.n_batchs))
+            self.log("Train perplexity: %g" % train_stats.ppl())
+            self.log("Train accuracy: %g" % train_stats.accuracy())
+            self.log("Sentences processed: %g" % train_stats.n_sents)
+            self.log(
+                "Average bsz: %4.0f/%4.0f/%2.0f"
+                % (
+                    train_stats.n_src_words / train_stats.n_batchs,
+                    train_stats.n_words / train_stats.n_batchs,
+                    train_stats.n_sents / train_stats.n_batchs,
+                )
+            )
 
-            self.maybe_log_tensorboard(train_stats,
-                                       "train",
-                                       lr,
-                                       patience,
-                                       step)
-
+            self.maybe_log_tensorboard(train_stats, "train", lr, patience, step)
         if valid_stats is not None:
-            self.log('Validation perplexity: %g' % valid_stats.ppl())
-            self.log('Validation accuracy: %g' % valid_stats.accuracy())
+            self.log("Validation perplexity: %g" % valid_stats.ppl())
+            self.log("Validation accuracy: %g" % valid_stats.accuracy())
 
-            self.maybe_log_tensorboard(valid_stats,
-                                       "valid",
-                                       lr,
-                                       patience,
-                                       step)
+            self.maybe_log_tensorboard(valid_stats, "valid", lr, patience, step)
