@@ -16,8 +16,9 @@ class Statistics(object):
     * elapsed time
     """
 
-    def __init__(self, loss=0, n_batchs=0, n_sents=0,
-                 n_words=0, n_correct=0, computed_metrics={}):
+    def __init__(
+        self, loss=0, n_batchs=0, n_sents=0, n_words=0, n_correct=0, computed_metrics={}
+    ):
         self.loss = loss
         self.n_batchs = n_batchs
         self.n_sents = n_sents
@@ -92,19 +93,19 @@ class Statistics(object):
             self.n_src_words += stat.n_src_words
 
     def accuracy(self):
-        """ compute accuracy """
+        """compute accuracy"""
         return 100 * (self.n_correct / self.n_words)
 
     def xent(self):
-        """ compute cross entropy """
+        """compute cross entropy"""
         return self.loss / self.n_words
 
     def ppl(self):
-        """ compute perplexity """
+        """compute perplexity"""
         return math.exp(min(self.loss / self.n_words, 100))
 
     def elapsed_time(self):
-        """ compute elapsed time """
+        """compute elapsed time"""
         return time.time() - self.start_time
 
     def output(self, step, num_steps, learning_rate, start):
@@ -120,27 +121,36 @@ class Statistics(object):
         if num_steps > 0:
             step_fmt = "%s/%5d" % (step_fmt, num_steps)
         logger.info(
-            ("Step %s; acc: %2.1f; ppl: %5.1f; xent: %2.1f; " +
-             "lr: %7.5f; sents: %7.0f; bsz: %4.0f/%4.0f/%2.0f; " +
-             "%3.0f/%3.0f tok/s; %6.0f sec;")
-            % (step_fmt,
-               self.accuracy(),
-               self.ppl(),
-               self.xent(),
-               learning_rate,
-               self.n_sents,
-               self.n_src_words / self.n_batchs,
-               self.n_words / self.n_batchs,
-               self.n_sents / self.n_batchs,
-               self.n_src_words / (t + 1e-5),
-               self.n_words / (t + 1e-5),
-               time.time() - start) +
-            "".join([" {}: {}".format(k, round(v, 2))
-                     for k, v in self.computed_metrics.items()]))
+            (
+                "Step %s; acc: %2.1f; ppl: %5.1f; xent: %2.1f; "
+                + "lr: %7.5f; sents: %7.0f; bsz: %4.0f/%4.0f/%2.0f; "
+                + "%3.0f/%3.0f tok/s; %6.0f sec;"
+            )
+            % (
+                step_fmt,
+                self.accuracy(),
+                self.ppl(),
+                self.xent(),
+                learning_rate,
+                self.n_sents,
+                self.n_src_words / self.n_batchs,
+                self.n_words / self.n_batchs,
+                self.n_sents / self.n_batchs,
+                self.n_src_words / (t + 1e-5),
+                self.n_words / (t + 1e-5),
+                time.time() - start,
+            )
+            + "".join(
+                [
+                    " {}: {}".format(k, round(v, 2))
+                    for k, v in self.computed_metrics.items()
+                ]
+            )
+        )
         sys.stdout.flush()
 
     def log_tensorboard(self, prefix, writer, learning_rate, patience, step):
-        """ display statistics to tensorboard """
+        """display statistics to tensorboard"""
         t = self.elapsed_time()
         writer.add_scalar(prefix + "/xent", self.xent(), step)
         writer.add_scalar(prefix + "/ppl", self.ppl(), step)
