@@ -16,10 +16,10 @@ from argparse import ArgumentParser
 
 parser = ArgumentParser()
 
-parser.add_argument("--nbest-hyp", type=str, help="file with nbest to rerank",
-                    required=True)
-parser.add_argument("--nbest-order", type=int, help="nbest order",
-                    required=True)
+parser.add_argument(
+    "--nbest-hyp", type=str, help="file with nbest to rerank", required=True
+)
+parser.add_argument("--nbest-order", type=int, help="nbest order", required=True)
 parser.add_argument("--output", type=str, help="output file", required=True)
 
 args = parser.parse_args()
@@ -28,7 +28,7 @@ args = parser.parse_args()
 def chunks(lgth, n):
     """Yield successive n-sized chunks from l."""
     for i in range(0, len(lgth), n):
-        yield lgth[i:i + n]
+        yield lgth[i : i + n]
 
 
 with codecs.open(args.nbest_hyp, encoding="utf-8") as file:
@@ -44,7 +44,9 @@ with codecs.open(args.output, "w", encoding="utf-8") as output_file:
         for hyp in nbest:
             bleu = 0
             for ref in nbest:
-                bleu = bleu + sacrebleu.sentence_bleu(hyp, [ref]).score / args.nbest_order
+                bleu = (
+                    bleu + sacrebleu.sentence_bleu(hyp, [ref]).score / args.nbest_order
+                )
                 texts.append(hyp)
                 scores.append(bleu)
         max_index = scores.index(max(scores))
